// src/app/utils/role.enum.ts
export enum Role {
    User = 'User',
    Admin = 'ADMIN',
    Quality = 'QUALITY_INSPECTOR',
    Superwise = 'SUPERVISOR',
    Maintenance = 'Maintenance',
    Operator = 'OPERATOR',
    Loading_Operateor = 'LOADING_OPERATOR',
    Unloading_Operateor = 'UNLOADING_OPERATOR',
    Monitor='MONITOR'

}